+++ packages/core/tests/cancelListing.test.ts
@@ -1,13 +1,13 @@
 import { describe, expect, it } from "vitest";
 
-import { accounts, config, mintERC721 } from "@ark-project/test";
+import { config, getAccounts, mintERC721 } from "@ark-project/test";
 
 import { cancelOrder, createListing } from "../src/actions/order/index.js";
 import { getOrderStatus } from "../src/actions/read/index.js";
 
 describe("cancelListing", () => {
   it("default", async () => {
-    const { seller, listingBroker } = accounts;
+    const { seller, listingBroker } = getAccounts();
     const { tokenId, tokenAddress } = await mintERC721({ account: seller });
 
     const { orderHash } = await createListing(config, {