+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ark-project/core
 
+## 3.0.0-beta.1
+
+### Patch Changes
+
+- 6d31848: Update contract address
+
+## 3.0.0-beta.0
+
+### Major Changes
+
+- Update SDKs
+
 ## 2.1.2
 
 ### Patch Changes